@@ -1,12 +1,12 @@
-package study.buddy.api;
+package studybuddy.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
+@SpringBootApplication
 public class StudyBuddyApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(StudyBuddyApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(StudyBuddyApplication.class, args);
+	}
 }
