@@ -1,9 +1,8 @@
-package study.buddy.api.endpoint;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+package studybuddy.api.endpoint;
 
 import lombok.extern.log4j.Log4j2;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @Log4j2
 @RestController
