@@ -1,4 +1,4 @@
-package study.buddy.api.user;
+package studybuddy.api.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
