@@ -1,12 +1,13 @@
-package study.buddy.api.user;
+package studybuddy.api.user;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
-import org.springframework.stereotype.Service;
-
 @Service
 public class UserService {
-
+    @Autowired
     private UserRepository userRepository;
 
     public Optional<User> findUser(Long userId) {
